Рефакторинг структуры основного сервиса
