+++ ewm-main-service/src/test/java/ru/practicum/controller/pub/PublicCategoriesControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.exception.NotFoundException;
-import ru.practicum.service.pub.PublicCategoriesServiceImpl;
+import ru.practicum.service.CategoriesService;
 
 import java.nio.charset.StandardCharsets;
 import java.util.List;
@@ -25,7 +25,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 class PublicCategoriesControllerTest {
     private final MockMvc mvc;
     @MockBean
-    private final PublicCategoriesServiceImpl service;
+    private final CategoriesService service;
 
     String baseUri = "/categories";
     String name = "name";
